requirements.txt
django==1.10.3
django-bower==5.2.0
django-jquery==3.1.1
django-pipeline==1.6.9
django_wysiwyg==0.8.0
django-ckeditor==5.2.1
xkcdpass==1.9.1
pycrypto==2.6.1
